feat: move biome_lsp/converters to a separate crate
