+++ crates/biome_lsp/Cargo.toml
@@ -14,23 +14,24 @@ version              = "0.0.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow              = { workspace = true }
-biome_analyze       = { workspace = true }
-biome_configuration = { workspace = true }
-biome_console       = { workspace = true }
-biome_deserialize   = { workspace = true }
-biome_diagnostics   = { workspace = true }
-biome_fs            = { workspace = true }
-biome_rowan         = { workspace = true }
-biome_service       = { workspace = true }
-biome_text_edit     = { workspace = true }
-futures             = "0.3.30"
-rustc-hash          = { workspace = true }
-serde               = { workspace = true, features = ["derive"] }
-serde_json          = { workspace = true }
-tokio               = { workspace = true, features = ["rt", "io-std"] }
-tower-lsp           = { version = "0.20.0" }
-tracing             = { workspace = true, features = ["attributes"] }
+anyhow               = { workspace = true }
+biome_analyze        = { workspace = true }
+biome_configuration  = { workspace = true }
+biome_console        = { workspace = true }
+biome_deserialize    = { workspace = true }
+biome_diagnostics    = { workspace = true }
+biome_fs             = { workspace = true }
+biome_lsp_converters = { workspace = true }
+biome_rowan          = { workspace = true }
+biome_service        = { workspace = true }
+biome_text_edit      = { workspace = true }
+futures              = "0.3.30"
+rustc-hash           = { workspace = true }
+serde                = { workspace = true, features = ["derive"] }
+serde_json           = { workspace = true }
+tokio                = { workspace = true, features = ["rt", "io-std"] }
+tower-lsp            = { version = "0.20.0" }
+tracing              = { workspace = true, features = ["attributes"] }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "macros"] }
