+++ crates/biome_lsp/src/capabilities.rs
@@ -1,5 +1,5 @@
-use crate::converters::{negotiated_encoding, PositionEncoding, WideEncoding};
 use biome_analyze::SUPPRESSION_ACTION_CATEGORY;
+use biome_lsp_converters::{negotiated_encoding, PositionEncoding, WideEncoding};
 use tower_lsp::lsp_types::{
     ClientCapabilities, CodeActionKind, CodeActionOptions, CodeActionProviderCapability,
     DocumentOnTypeFormattingOptions, OneOf, PositionEncodingKind, ServerCapabilities,
