+++ crates/biome_lsp/src/handlers/rename.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use crate::converters::from_proto;
+use biome_lsp_converters::from_proto;
 use crate::diagnostics::LspError;
 use crate::{session::Session, utils};
 use anyhow::{Context, Result};
