+++ crates/biome_lsp/src/lib.rs
@@ -1,5 +1,4 @@
 mod capabilities;
-mod converters;
 mod diagnostics;
 mod documents;
 mod extension_settings;
