+++ crates/biome_lsp/src/utils.rs
@@ -1,5 +1,3 @@
-use crate::converters::line_index::LineIndex;
-use crate::converters::{from_proto, to_proto, PositionEncoding};
 use anyhow::{ensure, Context, Result};
 use biome_analyze::ActionCategory;
 use biome_console::fmt::Termcolor;
@@ -9,6 +7,8 @@ use biome_diagnostics::termcolor::NoColor;
 use biome_diagnostics::{
     Applicability, {Diagnostic, DiagnosticTags, Location, PrintDescription, Severity, Visit},
 };
+use biome_lsp_converters::line_index::LineIndex;
+use biome_lsp_converters::{from_proto, to_proto, PositionEncoding};
 use biome_rowan::{TextRange, TextSize};
 use biome_service::workspace::CodeAction;
 use biome_text_edit::{CompressedOp, DiffOp, TextEdit};
@@ -380,8 +380,8 @@ pub(crate) fn apply_document_changes(
 #[cfg(test)]
 mod tests {
     use super::apply_document_changes;
-    use crate::converters::line_index::LineIndex;
-    use crate::converters::{PositionEncoding, WideEncoding};
+    use biome_lsp_converters::line_index::LineIndex;
+    use biome_lsp_converters::{PositionEncoding, WideEncoding};
     use biome_text_edit::TextEdit;
     use tower_lsp::lsp_types as lsp;
     use tower_lsp::lsp_types::{Position, Range, TextDocumentContentChangeEvent};
